pyproject.toml
[tool.poetry]
name = "ifn712"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = ">=3.2.2"
numpy = ">=1.18.5"
opencv-python = ">=4.1.1"
pillow = ">=7.1.2"
pyyaml = ">=5.3.1"
requests = ">=2.23.0"
scipy = ">=1.4.1"
torchvision = ">=0.8.1"
tqdm = ">=4.64.0"
tensorboard = ">=2.4.1"
pandas = ">=1.1.4"
seaborn = ">=0.11.0"
flask = "^2.2.2"
mediapipe-silicon = "^0.8.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
